bunch of updates
1. Remove the Pytorch availability checking
2. Separate cropping, pre-processing with feature extraction
    1) gray2rgb(), crop_patch() and crop_simple are moved to shared extract_features_utils.py
    2) torch_utils.py only serves as feature extraction
3. Remove test_gray() from unittest
