+++ spacer/config.py
@@ -127,7 +127,6 @@ MIN_TRAINIMAGES = 10
 
 # Check access to select which tests to run.
 HAS_CAFFE = importlib.util.find_spec("caffe") is not None
-HAS_TORCH = importlib.util.find_spec("torch") is not None
 
 conn = get_s3_conn()
 try:
